+++ src/app/features/loadouts/Loadout/GearList/Gear/LoadoutGear.js
@@ -2,7 +2,7 @@ import React, { useState } from 'react'
 import PropTypes from 'prop-types'
 
 import ConfirmDeleteDialog from 'app/shared/components/dialogs/ConfirmDeleteDialog'
-import LoadoutItem from 'app/shared/components/display/LoadoutItem'
+import LoadoutItem from 'app/shared/components/display/LoadoutItem/LoadoutItem'
 
 export default function LoadoutGear ({ gear, canDelete, onDelete }) {	
 	let [ isDialogOpen, setIsDialogOpen ] = useState(false)
