--- src/app/shared/components/cards/entities/index.js
@@ -1,2 +1,0 @@
-export { default as ArmoryCard } from './ArmoryCard'
-export { default as LoadoutCard } from './LoadoutCard'
