+++ src/app/shared/components/cards/index.js
@@ -1,2 +1,4 @@
 export { default as AddCard } from './AddCard'
 export { default as CardList } from './CardList'
+export { default as ArmoryCard } from './entities/ArmoryCard'
+export { default as LoadoutCard } from './entities/LoadoutCard'
