+++ src/app/shared/components/display/LoadoutItem/LoadoutItem.css
