+++ src/app/shared/components/display/LoadoutItem/LoadoutItem.js
@@ -1,8 +1,8 @@
 import React from 'react'
 import PropTypes from 'prop-types'
 
-import DeleteButton from '../buttons/DeleteButton'
-import LoadoutItemImage from '../images/LoadoutItemImage'
+import DeleteButton from '../../buttons/DeleteButton'
+import LoadoutItemImage from './LoadoutItemImage'
 
 import './LoadoutItem.css'
