+++ src/app/shared/components/display/LoadoutItem/index.js
@@ -1,0 +1,1 @@
+export { default } from './LoadoutItem'
